requirements.txt
# Core CV
ultralytics>=8.4.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# LLM providers
ollama>=0.4.0
openai>=1.30.0
google-genai>=1.0.0
anthropic>=0.30.0
httpx>=0.27.0

# Agentic orchestration
langgraph>=0.4.0
langchain-core>=0.3.0

# UI
streamlit>=1.45.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
